# [Sofa.Type] Change (?) default constructor for fixed_array